+++ src/core/server/systems/world.ts
@@ -1,4 +1,5 @@
 import * as alt from 'alt-server';
+import { WORLD_WEATHER } from '../../shared/enums/weather';
 import { DEFAULT_CONFIG } from '../athena/main';
 
 /* -- Top of Map --
@@ -22,10 +23,49 @@ const worldDivision = 6;
 const maxY = 8000;
 const minY = -4000;
 
+let weatherOverride = false;
+let weatherOverrideName = null;
+let timeOverride = false;
+let timeOverrideHour = null;
+
 export class World {
     static minMaxGroups: Array<{ minY: number; maxY: number }>;
     static hour: number = DEFAULT_CONFIG.BOOTUP_HOUR;
     static minute: number = DEFAULT_CONFIG.BOOTUP_MINUTE;
+
+    /**
+     * Used to override weather setting for all players.
+     * Automatically synchronized.
+     * @static
+     * @param {boolean} value
+     * @param {string} [_weatherName='']
+     * @memberof World
+     */
+    static setOverrideWeather(value: boolean, _weatherName: string = '') {
+        if (!_weatherName) {
+            _weatherName = WORLD_WEATHER.EXTRA_SUNNY;
+        }
+
+        weatherOverride = value;
+        weatherOverrideName = _weatherName;
+    }
+
+    /**
+     * Used to override time setting for all players.
+     * Automatically synchronized.
+     * @static
+     * @param {boolean} value
+     * @param {number} _hour
+     * @memberof World
+     */
+    static setOverrideTime(value: boolean, _hour: number) {
+        if (typeof _hour === 'string') {
+            _hour = parseInt(_hour);
+        }
+
+        timeOverride = value;
+        timeOverrideHour = _hour;
+    }
 
     /**
      * Generates a reference grid for weather and objects.
@@ -40,7 +80,7 @@ export class World {
         for (let i = 0; i < division; i++) {
             const result = {
                 maxY: maxY - (total / division) * i,
-                minY: maxY - 2000 - (total / division) * i
+                minY: maxY - 2000 - (total / division) * i,
             };
 
             groups.push(result);
@@ -83,14 +123,30 @@ export class World {
 
     static getGridSpace(player: alt.Player): number {
         const gridSpace = World.minMaxGroups.findIndex(
-            (pos) => player && player.valid && player.pos.y > pos.minY && player.pos.y < pos.maxY
+            (pos) => player && player.valid && player.pos.y > pos.minY && player.pos.y < pos.maxY,
         );
 
         return gridSpace === -1 ? 0 : gridSpace;
     }
 
     static getWeatherByGrid(gridIndex: number): string {
+        if (weatherOverride) {
+            return weatherOverrideName;
+        }
+
         return DEFAULT_CONFIG.WEATHER_ROTATION[gridIndex];
+    }
+
+    static getWorldHour() {
+        if (timeOverride) {
+            return timeOverrideHour;
+        }
+
+        return World.hour;
+    }
+
+    static getWorldMinute() {
+        return World.minute;
     }
 }